final changes to the controllers
